# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GlobeTrotter API
  description: >
    Public API surface exposed by the GlobeTrotter API Gateway.
    Every path below is served by exactly one microservice behind the
    gateway (see api-gateway/README.md for the routing table) but the
    frontend only ever needs to know this single contract.
  version: "1.0.0"
  contact:
    name: GlobeTrotter Backend
servers:
  - url: http://localhost:5000
    description: Local docker-compose
  - url: https://api.globetrotter.example.com
    description: Production (via Ingress)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
        data: {}
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new visitor account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string}
                email: {type: string}
                preferences:
                  type: array
                  items: {type: string}
      responses:
        "201": {description: Created, content: {application/json: {schema: {$ref: "#/components/schemas/SuccessResponse"}}}}
        "400": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
        "409": {description: Username already exists}
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in as a visitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        "200": {description: OK -- returns access + refresh JWT}
        "401": {description: Invalid credentials}
  /api/auth/admin/login:
    post:
      tags: [Auth]
      summary: Log in as the administrator (constant credentials)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string, description: "Admin email"}
                password: {type: string}
      responses:
        "200": {description: OK}
        "401": {description: Invalid credentials}
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        "200": {description: OK}
        "401": {description: Invalid or expired refresh token}
  /api/auth/verify:
    get:
      tags: [Auth]
      summary: Validate a token
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Token is valid}
        "401": {description: Invalid or missing token}
...